Port pspgl to PSPSDK.
Remove the sce* headers and stubs as part of the port to PSPSDK.

Move module-specific initialization code out of pspgl and into the samples.
 
Fixup floating point constants in the samples, and add
-fsingle-precision-constant to the pspgl Makefile to remove doubles.

Makefile: Add -G0 to CFLAGS to disable the $gp register.  Put autogenerated
dependencies in a seperate subdirectory to reduce clutter.  Create GL if it
doesn't exist (need to find a way to do this on platforms that don't support
symbolic links).

+++ pspgl_vidmem.c
@@ -1,8 +1,8 @@
 #include <stdlib.h>
 #include <string.h>
+#include <pspdisplay.h>
+#include <pspge.h>
 #include "pspgl_internal.h"
-#include "sceDisplay.h"
-#include "sceGe.h"
 
 
 /* PSP pixelformats */
@@ -52,7 +52,7 @@ void *vidmem_map_insert_new (unsigned long idx, unsigned long adr, unsigned long
 
 void* pspgl_vidmem_alloc (unsigned long size)
 {
-	unsigned long start = sceGeEdramGetAddr();
+	unsigned long start = (unsigned long) sceGeEdramGetAddr();
 	unsigned long adr = start;
 	unsigned long i;
